requirements.txt
flask
langchain
langchain-core
langchain-community
langchain-groq
wikipedia
duckduckgo-search
ddgs
python-dotenv
pydantic